# core-common-kycb-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-common-kycb-mgmt
  description: KYC/B & AML Management Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/kycb
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/identity/parties/{partyId}/kyb:
    get:
      tags:
        - KYB Verification
      summary: List KYB verifications for a party
      description: Retrieves all KYB verifications for the specified party ID with optional filtering
      operationId: listKybVerifications
      parameters:
        - name: partyId
          in: path
          description: ID of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.partyId
          in: query
          description: Exact filter for party id
          required: false
          schema:
            type: string
        - name: filters.verificationStatus
          in: query
          description: Exact filter for verification status
          required: false
          schema:
            type: string
        - name: filters.verificationDate
          in: query
          description: Exact filter for verification date
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[verificationDate].from
          in: query
          description: Filter verification date from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[verificationDate].to
          in: query
          description: Filter verification date to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.mercantileRegistryVerified
          in: query
          description: Exact filter for mercantile registry verified
          required: false
          schema:
            type: boolean
        - name: filters.deedOfIncorporationVerified
          in: query
          description: Exact filter for deed of incorporation verified
          required: false
          schema:
            type: boolean
        - name: filters.businessStructureVerified
          in: query
          description: Exact filter for business structure verified
          required: false
          schema:
            type: boolean
        - name: filters.uboVerified
          in: query
          description: Exact filter for ubo verified
          required: false
          schema:
            type: boolean
        - name: filters.taxIdVerified
          in: query
          description: Exact filter for tax id verified
          required: false
          schema:
            type: boolean
        - name: filters.operatingLicenseVerified
          in: query
          description: Exact filter for operating license verified
          required: false
          schema:
            type: boolean
        - name: filters.verificationNotes
          in: query
          description: Exact filter for verification notes
          required: false
          schema:
            type: string
        - name: filters.riskScore
          in: query
          description: Exact filter for risk score
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[riskScore].from
          in: query
          description: Filter risk score from value
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[riskScore].to
          in: query
          description: Filter risk score to value
          required: false
          schema:
            type: integer
            format: int32
        - name: filters.riskLevel
          in: query
          description: Exact filter for risk level
          required: false
          schema:
            type: string
        - name: filters.nextReviewDate
          in: query
          description: Exact filter for next review date
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[nextReviewDate].from
          in: query
          description: Filter next review date from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[nextReviewDate].to
          in: query
          description: Filter next review date to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.dateCreated
          in: query
          description: Exact filter for date created
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateCreated].from
          in: query
          description: Filter date created from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateCreated].to
          in: query
          description: Filter date created to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.dateUpdated
          in: query
          description: Exact filter for date updated
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateUpdated].from
          in: query
          description: Filter date updated from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateUpdated].to
          in: query
          description: Filter date updated to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Successfully retrieved KYB verifications
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponse'
    post:
      tags:
        - KYB Verification
      summary: Start new KYB verification
      description: Creates a new KYB verification for the specified party
      operationId: createKybVerification
      parameters:
        - name: partyId
          in: path
          description: ID of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KybVerificationDTO'
        required: true
      responses:
        '201':
          description: Successfully created KYB verification
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/KybVerificationDTO'
...
  /api/v1/identity/parties/{partyId}/kyb/{verificationId}:
    get:
      tags:
        - KYB Verification
      summary: Get specific KYB verification
      description: Retrieves a specific KYB verification by its ID
      operationId: getKybVerification
      parameters:
        - name: partyId
          in: path
          description: ID of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: verificationId
          in: path
          description: ID of the verification
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved KYB verification
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/KybVerificationDTO'
        '404':
          description: KYB verification not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/KybVerificationDTO'
    delete:
      tags:
        - KYB Verification
      summary: Delete KYB verification
      description: Deletes a KYB verification
      operationId: deleteKybVerification
      parameters:
        - name: partyId
          in: path
          description: ID of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: verificationId
          in: path
          description: ID of the verification
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Successfully deleted KYB verification
        '404':
          description: KYB verification not found
    patch:
      tags:
        - KYB Verification
      summary: Update KYB verification
      description: Updates an existing KYB verification
      operationId: updateKybVerification
      parameters:
        - name: partyId
          in: path
          description: ID of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: verificationId
          in: path
          description: ID of the verification
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KybVerificationDTO'
        required: true
      responses:
        '200':
          description: Successfully updated KYB verification
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/KybVerificationDTO'
        '404':
          description: KYB verification not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/KybVerificationDTO'
...
  /api/v1/documents/verification/{documentId}:
    get:
      tags:
        - Verification Documents
      summary: Get verification document
      description: Retrieves a specific verification document by its ID
      operationId: getVerificationDocument_1
      parameters:
        - name: documentId
          in: path
          description: ID of the document
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved verification document
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VerificationDocumentDTO'
        '404':
          description: Verification document not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VerificationDocumentDTO'
    delete:
      tags:
        - Verification Documents
      summary: Delete verification document
      description: Deletes a verification document
      operationId: deleteVerificationDocument_1
      parameters:
        - name: documentId
          in: path
          description: ID of the document
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Successfully deleted verification document
        '404':
          description: Verification document not found
    patch:
      tags:
        - Verification Documents
      summary: Update verification document
      description: Updates an existing verification document
      operationId: updateVerificationDocument_1
      parameters:
        - name: documentId
          in: path
          description: ID of the document
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerificationDocumentDTO'
        required: true
      responses:
        '200':
          description: Successfully updated verification document
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VerificationDocumentDTO'
        '404':
          description: Verification document not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VerificationDocumentDTO'
...
components:
  schemas:
...
    VerificationDocumentDTO:
      required:
        - documentReference
        - documentType
        - identityDocumentId
        - isVerified
        - kycVerificationId
        - verificationPurpose
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        verificationDocumentId:
          type: string
          format: uuid
          readOnly: true
        kycVerificationId:
          type: string
          format: uuid
        identityDocumentId:
          type: string
          format: uuid
        documentType:
          maxLength: 50
          minLength: 0
          type: string
        verificationPurpose:
          maxLength: 50
          minLength: 0
          type: string
        documentReference:
          maxLength: 100
          minLength: 0
          type: string
        documentSystemId:
          maxLength: 100
          minLength: 0
          type: string
        isVerified:
          type: boolean
        verificationNotes:
          maxLength: 1000
          minLength: 0
          type: string
        expiryDate:
          type: string
          format: date-time
    KybVerificationDTO:
      required:
        - businessStructureVerified
        - deedOfIncorporationVerified
        - mercantileRegistryVerified
        - operatingLicenseVerified
        - partyId
        - riskLevel
        - riskScore
        - taxIdVerified
        - uboVerified
        - verificationDate
        - verificationStatus
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        kybVerificationId:
          type: string
          format: uuid
          readOnly: true
        partyId:
          type: string
          format: uuid
        verificationStatus:
          maxLength: 50
          minLength: 0
          type: string
        verificationDate:
          type: string
          format: date-time
        mercantileRegistryVerified:
          type: boolean
        deedOfIncorporationVerified:
          type: boolean
        businessStructureVerified:
          type: boolean
        uboVerified:
          type: boolean
        taxIdVerified:
          type: boolean
        operatingLicenseVerified:
          type: boolean
        verificationNotes:
          maxLength: 1000
          minLength: 0
          type: string
        riskScore:
          maximum: 100
          minimum: 0
          type: integer
          format: int32
        riskLevel:
          maxLength: 20
          minLength: 0
          type: string
        nextReviewDate:
          type: string
          format: date-time
...
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.